Trailconn refactor
