+++ contracts/trail.service/trail.service.hpp
@@ -1,6 +1,13 @@
+/**
+ * 
+ * 
+ */
+
 #pragma once
 
-#include <trail.connections/trailconn.voting.hpp> //Import trailservice voting data definitions
+#include <trail.connections/trailconn.voting.hpp>
+#include <trail.connections/trailconn.system.hpp>
+#include <trail.connections/trailconn.tokens.hpp>
 
 #include <eosiolib/asset.hpp>
 #include <eosiolib/eosio.hpp>
@@ -10,8 +17,8 @@
 using namespace eosio;
 
 class trail : public contract {
+    
     public:
-
         /**
          * Constructor sets environment singleton upon contract deployment, and gets environment for every action call.
         */
@@ -76,17 +83,6 @@ class trail : public contract {
         void unregballot(account_name publisher);
 
     protected:
-
-        /// @abi table registries i64
-        struct registration {
-            asset native;
-            account_name publisher;
-
-            uint64_t primary_key() const { return native.symbol.name(); }
-            uint64_t by_publisher() const { return publisher; }
-        };
-
-        typedef multi_index<N(registries), registration> registries_table;
 
         environment_singleton env_singleton;
         environment env_struct;
